refactor: extract `getOptionsFromCli`
